// src/types/colorType.ts
export type StatColors = {
    FOR: string;
    DEX: string;
    CON: string;
    INT: string;
    SAG: string;
    CHA: string;
  
    proficiencyField: string;
    inspirationField: string;
    passivePercField: string;
    initiativeField: string;
    speedField: string;
    ddSaveField: string;
};